feat: rework rendering to remove use of ice edge pieces. add ground on all boundaries. rework levels to have correct padding'

+++ assemble_level.c
@@ -63,12 +63,12 @@ int main(int argc, char** argv) {
     for (int i=0; i<TOTAL_NUMBER; i++) {
         // w h tx1 ty1 tx2 ty2
         stbrp_rect rect = rects[i];
-        float x1 = 1.0 * rect.x / total_width;
+        double x1 = 1.0 * rect.x / total_width;
         // this is because data is saved upside down compared to opengl loader.
-        float y1 = 1.0 * (total_height - (rect.y + rect.h)) / total_height;
-        float x2 = 1.0 * (rect.x + rect.w) / total_width;
-        float y2 = 1.0 * (total_height-rect.y) / total_height;
-        fprintf(sprite_data, "%i %i %f %f %f %f\n", rect.w, rect.h, x1, y1, x2, y2);
+        double y1 = 1.0 * (total_height - (rect.y + rect.h)) / total_height;
+        double x2 = 1.0 * (rect.x + rect.w) / total_width;
+        double y2 = 1.0 * (total_height-rect.y) / total_height;
+        fprintf(sprite_data, "%i %i %lf %lf %lf %lf\n", rect.w, rect.h, x1, y1, x2, y2);
     }
     fclose(sprite_data);
